requirements.txt
pyyaml
openpyxl
plotly
numpy
pandas
python-docx
streamlit
